refactor: Repository를 인터페이스로 분리
- 새로운 Repository 구현체를 추가하기 위해 분리함
