+++ src/main/java/com/wootecam/jspcafe/domain/QuestionRepository.java
@@ -1,0 +1,13 @@
+package com.wootecam.jspcafe.domain;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface QuestionRepository {
+
+    void save(final Question question);
+
+    List<Question> findAll();
+
+    Optional<Question> findById(final Long id);
+}
