+++ src/main/java/com/wootecam/jspcafe/repository/InMemoryUserRepository.java
@@ -1,6 +1,7 @@
 package com.wootecam.jspcafe.repository;
 
 import com.wootecam.jspcafe.domain.User;
+import com.wootecam.jspcafe.domain.UserRepository;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -8,7 +9,7 @@ import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicLong;
 
-public class UserRepository {
+public class InMemoryUserRepository implements UserRepository {
 
     private final Map<Long, User> store = new ConcurrentHashMap<>();
     private final AtomicLong idGenerator = new AtomicLong(1);
@@ -17,18 +18,22 @@ public class UserRepository {
         return idGenerator.getAndIncrement();
     }
 
+    @Override
     public void save(User user) {
         store.put(user.getId(), user);
     }
 
+    @Override
     public List<User> findAll() {
         return new ArrayList<>(store.values());
     }
 
+    @Override
     public Optional<User> findById(final Long id) {
         return Optional.ofNullable(store.get(id));
     }
 
+    @Override
     public void update(final User user) {
         store.put(user.getId(), user);
     }
